requirements.txt
WTForms==3.0.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
SQLAlchemy==2.0.7
bcrypt==4.0.1
email-validator==1.3.1
gunicorn==20.1.0
length==1.0.1
Pillow==9.5.0
pip==22.3.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
zipp==3.15.0
psycopg2==2.9.6